pyproject.toml
[project]
name = "drf-jwt-stateless"
version = "0.1.0"
description = ""
authors = [
    {name = "Iago Corrêa",email = "iago.cunhac@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "django-redis (>=6.0.0,<7.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
